Fix #18 make Thoth.Json.Net reference min Newtonsoft.Json. version.
Support this via the use of Paket's lowest_matching setting at the
package level in paket.dependencies.  This results in Thoth getting a
reference to the the min version, but Paket.pack still creates a min
inclusive dependency version in the nuspec (== in paket.dependencies
would have specified an exact verion which would not be correct).