Re-organize the API
I wasn't happy with exposing the "LowLevel" API near the standard API. Reason was the LowLevel API can have unexpected behaviour as we can't garantee that the provided `System.Type` is indeed representing `'T`.

A solution could have been to use `obj` as the output and remove the generic type but then the user need to cast the himself the obj. I think using the generic type is more user friendly for now.